Integrated PuzzleSeg detection. Review of height/workspace segmentation.
The depth values vary so much that it is difficult to actually get a goot reading.  Need to rely on color mask more than depth for getting the glove.  The right code for doing so has to be established. Will most likely be some kind of grow from joint binary mask.  Can it be done with watershed?